Move reductions out of array_ops and into array_reduce. And move DG-aware reductions into array_dg_reduce. All unit tests pass on CPUs.
