+++ zero/array_ops.c
@@ -1,15 +1,14 @@
-#include "gkyl_range.h"
+#include <gkyl_array_ops.h>
+#include <gkyl_array_ops_priv.h>
+#include <gkyl_alloc_flags_priv.h>
+#include <gkyl_range.h>
+#include <gkyl_util.h>
+
 #include <assert.h>
 #include <float.h>
 #include <math.h>
 #include <stdio.h>
 #include <string.h>
-
-#include <gkyl_util.h>
-#include <gkyl_array_ops.h>
-#include <gkyl_array_ops_priv.h>
-#include <gkyl_array_reduce.h>
-#include <gkyl_alloc_flags_priv.h>
 
 bool
 gkyl_array_copy_func_is_cu_dev(const struct gkyl_array_copy_func *bc)
@@ -166,127 +165,6 @@ gkyl_array_shiftc(struct gkyl_array* out, double a, unsigned k)
   for (size_t i=0; i<out->size; ++i)
     out_d[i*NCOM(out)+k] = a+out_d[i*NCOM(out)+k];
   return out;
-}
-
-void 
-gkyl_array_reduce(double *out, const struct gkyl_array *arr, enum gkyl_array_op op)
-{
-  assert(arr->type == GKYL_DOUBLE);
-
-#ifdef GKYL_HAVE_CUDA
-  if (gkyl_array_is_cu_dev(arr)) {
-    switch (op) {
-      case GKYL_MAX:
-        gkyl_array_reduce_max_cu(out, arr);
-        break;
-      case GKYL_MIN:
-        gkyl_array_reduce_min_cu(out, arr);
-        break;
-      case GKYL_SUM:
-        gkyl_array_reduce_sum_cu(out, arr);
-        break;
-    }
-    return;
-  }
-#endif
-
-  long nc = NCOM(arr);
-  double *arr_d = arr->data;
-
-  switch (op) {
-    case GKYL_MIN:
-      for (long k=0; k<nc; ++k) out[k] = DBL_MAX;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        for (long k=0; k<nc; ++k)
-          out[k] = fmin(out[k], d[k]);
-      }
-      break;
-
-    case GKYL_MAX:
-      for (long k=0; k<nc; ++k) out[k] = -DBL_MAX;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        for (long k=0; k<nc; ++k)
-          out[k] = fmax(out[k], d[k]);
-      }
-      break;
-
-    case GKYL_SUM:
-      for (long k=0; k<nc; ++k) out[k] = 0;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        for (long k=0; k<nc; ++k)
-          out[k] += d[k];
-      }
-      break;
-  }
-}
-
-void 
-gkyl_array_reducec_dg(double *out, const struct gkyl_array *arr, int comp,
-  enum gkyl_array_op op, const struct gkyl_basis *basis)
-{
-  assert(arr->type == GKYL_DOUBLE);
-
-#ifdef GKYL_HAVE_CUDA
-  if (gkyl_array_is_cu_dev(arr)) {
-    switch (op) {
-      case GKYL_MAX:
-        gkyl_array_reducec_dg_max_cu(out, arr, comp, basis);
-        break;                                            
-      case GKYL_MIN:                                      
-        gkyl_array_reducec_dg_min_cu(out, arr, comp, basis);
-        break;                                            
-      case GKYL_SUM:                                      
-        gkyl_array_reducec_dg_sum_cu(out, arr, comp, basis);
-        break;
-    }
-    return;
-  }
-#endif
-
-  double *arr_d = arr->data;
-
-  int num_nodes = pow(basis->poly_order+1,basis->ndim);
-
-  switch (op) {
-    case GKYL_MIN:
-      out[0] = DBL_MAX;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] = fmin(out[0], arr_nodal[k]);
-        }
-      }
-      break;
-
-    case GKYL_MAX:
-      out[0] = -DBL_MAX;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] = fmax(out[0], arr_nodal[k]);
-        }
-      }
-      break;
-
-    case GKYL_SUM:
-      out[0] = 0;
-      for (size_t i=0; i<arr->size; ++i) {
-        const double *d = gkyl_array_cfetch(arr, i);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] += arr_nodal[k];
-        }
-      }
-      break;
-  }
 }
 
 // range based methods
@@ -509,140 +387,6 @@ gkyl_array_shiftc_range(struct gkyl_array* out, double a, unsigned k, const stru
   }
   return out;
 }
-
-void
-gkyl_array_reduce_range(double *res,
-  const struct gkyl_array *arr, enum gkyl_array_op op, const struct gkyl_range *range)
-{
-  assert(arr->type == GKYL_DOUBLE);
-
-#ifdef GKYL_HAVE_CUDA
-  if (gkyl_array_is_cu_dev(arr)) {
-    switch (op) {
-      case GKYL_MAX:
-        gkyl_array_reduce_range_max_cu(res, arr, range);
-        break;
-      case GKYL_MIN:
-        gkyl_array_reduce_range_min_cu(res, arr, range);
-        break;
-      case GKYL_SUM:
-        gkyl_array_reduce_range_sum_cu(res, arr, range);
-        break;
-    }
-    return;
-  }
-#endif
-
-  long n = NCOM(arr);
-  struct gkyl_range_iter iter;
-  gkyl_range_iter_init(&iter, range);
-
-  switch (op) {
-    case GKYL_MIN:
-      for (long i=0; i<n; ++i) res[i] = DBL_MAX;
-
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        for (long i=0; i<n; ++i)
-          res[i] = fmin(res[i], d[i]);
-      }
-      break;
-    case GKYL_MAX:
-      for (long i=0; i<n; ++i) res[i] = -DBL_MAX;
-
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        for (long i=0; i<n; ++i)
-          res[i] = fmax(res[i], d[i]);
-      }
-      break;
-    case GKYL_SUM:
-      for (long i=0; i<n; ++i) res[i] = 0;
-
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        for (long i=0; i<n; ++i)
-          res[i] += d[i];
-      }
-      break;
-  }
-}
-
-void 
-gkyl_array_reducec_dg_range(double *out, const struct gkyl_array *arr, int comp,
-  enum gkyl_array_op op, const struct gkyl_basis *basis, const struct gkyl_range *range)
-{
-  assert(arr->type == GKYL_DOUBLE);
-
-#ifdef GKYL_HAVE_CUDA
-  if (gkyl_array_is_cu_dev(arr)) {
-    switch (op) {
-      case GKYL_MAX:
-        gkyl_array_reducec_dg_range_max_cu(out, arr, comp, basis, range);
-        break;
-      case GKYL_MIN:
-        gkyl_array_reducec_dg_range_min_cu(out, arr, comp, basis, range);
-        break;
-      case GKYL_SUM:
-        gkyl_array_reducec_dg_range_sum_cu(out, arr, comp, basis, range);
-        break;
-    }
-    return;
-  }
-#endif
-
-  double *arr_d = arr->data;
-
-  int num_nodes = pow(basis->poly_order+1,basis->ndim);
-
-  struct gkyl_range_iter iter;
-  gkyl_range_iter_init(&iter, range);
-
-  switch (op) {
-    case GKYL_MIN:
-      out[0] = DBL_MAX;
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] = fmin(out[0], arr_nodal[k]);
-        }
-      }
-      break;
-
-    case GKYL_MAX:
-      out[0] = -DBL_MAX;
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] = fmax(out[0], arr_nodal[k]);
-        }
-      }
-      break;
-
-    case GKYL_SUM:
-      out[0] = 0;
-      while (gkyl_range_iter_next(&iter)) {
-        long start = gkyl_range_idx(range, iter.idx);
-        const double *d = gkyl_array_cfetch(arr, start);
-        double arr_nodal[num_nodes];
-        for (int k=0; k<num_nodes; ++k) {
-          basis->modal_to_quad_nodal(&d[comp*basis->num_basis], arr_nodal, k);
-          out[0] += arr_nodal[k];
-        }
-      }
-      break;
-  }
-}
-
 
 struct gkyl_array*
 gkyl_array_copy_range(struct gkyl_array *out,
