+++ zero/gkyl_array_ops.h
@@ -256,60 +256,6 @@ struct gkyl_array* gkyl_array_copy_range_to_range(struct gkyl_array *out,
   const struct gkyl_array *inp, const struct gkyl_range *out_range, const struct gkyl_range *inp_range);
 
 /**
- * Perform an "reduce" operation of data in the array.
- *
- * @param res On output, reduces values (ncomp size).
- * @param arr Array to perform reduction on.
- * @param op Reduction operators.
- */
-void gkyl_array_reduce(double *res, const struct gkyl_array *arr, enum gkyl_array_op op);
-
-/**
- * Perform a "reduce" operation of data in the array accounting for the DG
- * representation within the array. It evaluates the DG field at Gauss-Legendre
- * nodes in each cell, and reduces them before reducing over cells. It allows
- * specification of the vector component for arrays that contain multiple
- * scalar DG fields.
- *
- * @param out Reduced output value (size 1).
- * @param arr Array to perform reduction on.
- * @param comp Vector component in arr to reduce.
- * @param op Reduction operators.
- * @param basis Basis DG coefficients are expanded in (device pointer if use_gpu=true).
- */
-void gkyl_array_reducec_dg(double *out, const struct gkyl_array *arr, int comp,
-  enum gkyl_array_op op, const struct gkyl_basis *basis);
-
-/**
- * Perform an "reduce" operation of data in the array. Data is reduced
- * component-wise.
- *
- * @param res On output, reduced values (ncomp size).
- * @param arr Array to perform reduction on.
- * @param op Reduction operators.
- * @param range Range specifying region.
- */
-void gkyl_array_reduce_range(double *res,
-  const struct gkyl_array *arr, enum gkyl_array_op op, const struct gkyl_range *range);
-
-/**
- * Perform a "reduce" operation of data in the array accounting for the DG
- * representation within the array. It evaluates the DG field at Gauss-Legendre
- * nodes in each cell, and reduces them before reducing over cells. It allows
- * specification of the vector component for arrays that contain multiple
- * scalar DG fields.
- *
- * @param out Reduced output value (size 1).
- * @param arr Array to perform reduction on.
- * @param comp Vector component in arr to reduce.
- * @param op Reduction operators.
- * @param basis Basis DG coefficients are expanded in (device pointer if use_gpu=true).
- * @param range Range specifying region.
- */
-void gkyl_array_reducec_dg_range(double *out, const struct gkyl_array *arr, int comp,
-  enum gkyl_array_op op, const struct gkyl_basis *basis, const struct gkyl_range *range);
-
-/**
  * Copy region of array into a buffer. The buffer must be preallocated
  * and at least of size arr->size*arr->elemSz bytes.
  *
